requirements_keras.txt
bert-for-tf2==0.14.4
jupyter==1.0.0
pandas==1.0.4
scikit-learn==0.23.1
tensorflow==2.2.0
